refactor(parser): move PatternParser to a new module and update imports
