+++ src/imgtools/dicom/sort/__init__.py
@@ -83,7 +83,7 @@ from imgtools.dicom.sort.exceptions import (
     SorterBaseError,
 )
 from imgtools.dicom.sort.highlighter import TagHighlighter
-from imgtools.dicom.sort.parser import PatternParser
+from imgtools.pattern_parser.parser import PatternParser
 from imgtools.dicom.sort.sort_method import FileAction, handle_file
 from imgtools.dicom.sort.sorter_base import SorterBase, resolve_path
 from imgtools.dicom.sort.utils import read_tags
