+++ src/imgtools/pattern_parser/parser.py
