+++ src/imgtools/pattern_parser/pattern_resolver.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, field
 from typing import Any, ClassVar, Dict, Tuple
 
 from imgtools.dicom.sort.exceptions import InvalidPatternError
-from imgtools.dicom.sort.parser import PatternParser
+from imgtools.pattern_parser.parser import PatternParser
 from imgtools.logging import logger
